Add swupdate configuration.
The following patch adds SWUpdate configuration and scripts as well as SoMLabs logos.

To use it, build and write console-image to SDcard. Then build update-image recipe.
It creates an update image (*.swu), which can be uploaded to the running board via web interface:
http://<board ip>:8080/

SWUpdate is configured to use two partition layout (rootfs, and update).
Bootloader is switched to new partition after a sucessfull image
download and installation.

Please note, the setup is tested on console-image using a SDcard version
if the SoM.

Signed-off-by: Jakub Turek <j_turek@elsta.pl>